create API, check DB

+++ internal/domain/entity/user.go
@@ -6,7 +6,7 @@ import "go.mongodb.org/mongo-driver/bson/primitive"
 
 type User struct {
 	Id    primitive.ObjectID `json:"_id,omitempty" bson:"_id,omitempty"`
-	Role  uint8              `json:"role"`
+	Role  int                `json:"role"`
 	Name  string             `json:"name"`
 	Email string             `json:"email"`
 }
